Replace old connection with gun connection
