+++ apps/discordbot/lib/discordbot/gateway/supervisor.ex
@@ -5,7 +5,7 @@ defmodule DiscordBot.Gateway.Supervisor do
   require Logger
 
   alias DiscordBot.Broker.Shovel
-  alias DiscordBot.Gateway.{Authenticator, Connection, GunConnection, Heartbeat}
+  alias DiscordBot.Gateway.{Authenticator, Connection, Heartbeat}
 
   def start_link(opts) do
     token = Keyword.fetch!(opts, :token)
@@ -66,8 +66,7 @@ defmodule DiscordBot.Gateway.Supervisor do
         id: :authenticator,
         restart: :transient
       ),
-      {Connection, url: url, token: token, broker: instance_broker},
-      {GunConnection, url: url, token: token, broker: instance_broker}
+      {Connection, url: url, token: token, broker: instance_broker}
     ]
 
     Supervisor.init(children, strategy: :one_for_all)
